Try to fix dbg to file
